Demo/task 01 1.5 b0

+++ user-service/src/main/java/com/plagod/service/UserManageService.java
@@ -11,7 +11,10 @@ import com.plagod.vo.user.UserRoleSnapshotVO;
 import java.util.List;
 
 public interface UserManageService {
-    UserPageResult pageUsers(long current, long size, String keyword);
+    UserPageResult pageUsers(
+            Integer current,
+            Integer size,
+            String keyword);
 
     UserVO getUser(Long userId);
