+++ user-service/src/main/java/com/plagod/service/UserOperationRequestService.java
@@ -4,7 +4,10 @@ import com.plagod.vo.user.UserOperationRequestPageResult;
 import com.plagod.dto.user.UserOperationReviewDTO;
 
 public interface UserOperationRequestService {
-    UserOperationRequestPageResult pageRequests(long current, long size, Integer status);
+    UserOperationRequestPageResult pageRequests(
+            Integer current,
+            Integer size,
+            Integer status);
 
     Long requestPurge(Long targetUserId, Long requesterId, String requesterName, String reason);
